requirements.txt
rplidar
# smbus
python-socketio
eventlet

# pip install -r requirements.txt -v